+++ docs/changelog.md
@@ -16,6 +16,7 @@ Ends September 16, 2020
 - add audit script.
 - add styles to Notifications.
 - add sh directory.
+- improve audit scripts.
 
 ## Sprint 3
 Ends August 31, 2020